classifier dict; add SCN result directly to query annData
