+++ pySingleCellNet/utils.py
@@ -19,7 +19,8 @@ def ctRename(sampTab, annCol, oldName, newName):
 #adapted from Sam's code
 def splitCommonAnnData(adata, ncells, dLevel="cell_ontology_class", cellid = None, cells_reserved = 3):
     if cellid == None: 
-        adata.obs[cellid] = adata.obs.index
+         adata.obs["cellid"] = adata.obs.index
+         cellid = "cellid"
     cts = set(adata.obs[dLevel])
 
     trainingids = np.empty(0)
